SP2-1837 Compile the list of previous versions for a Sentence Plan on the AAP

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/arnscoordinatorapi/integrations/plan/api/SentencePlanApi.kt
@@ -183,7 +183,7 @@ class SentencePlanApi(
         list.map {
           VersionDetails(
             it.uuid,
-            it.version,
+            it.version.toLong(),
             it.createdAt,
             it.updatedAt,
             it.status,
